requirements.txt
flask>=2.3.0
crawl4ai>=0.7.0
transformers>=4.30.0
torch>=2.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
lxml>=4.9.0
selenium>=4.10.0
webdriver-manager>=4.0.0 